Crate encrypted_package

A library that helps you send data in encrypted form.

Features

async

Adds asend and areceive for asynchronous sending and receiving.

proof-of-work

Adds the ability to generate and check proof of work.

Depending on the difficulty of the proof, it takes time. It is better to use this in conjunction with async blocking code spawner.

In addition, it increases the size of data transmitted by 8 bytes.

tracing

Adds basic tracing with instrumentation. Example how to create a subscriber:

let subscriber = tracing_subscriber::Registry::default()
	.with(tracing_subscriber::EnvFilter::new("INFO"))
	.with(tracing_bunyan_formatter::JsonStorageLayer)
	.with(tracing_bunyan_formatter::BunyanFormattingLayer::new(
		"test".to_owned(),
		std::io::stdout,
	));
tracing::subscriber::set_global_default(subscriber).context("Failed to set.")

Algorithm

First it sends a data with a size of 8 bytes, which contains the size of the encrypted package. Then it sends the encrypted package bytes.

Specifications

Encryption: RSA-PKCS1-OAEP, AES-GCM-256

Signing: RSA-PKCS1-PSS

Hashing: SHA-256

Example (synchronous)

Sending Package:

let config = encrypted_package::DefaultConfig::try_default()?;
let mut package = encrypted_package::Package::new(
	&recipient_public_key, None::<Vec<u8>>, vec![1, 2, 3], config,
)?;
// If "proof-of-work" feature is enabled
// package.generate_proof_of_work();
package.sign(&sender_private_key)?;
package.send(&mut stream, None)?;

Receiving Package (without decrypting):

let package = encrypted_package::Package::<
	Vec<u8>, Vec<u8>, encrypted_package::DefaultConfig,
>::receive(&mut stream, None)?;
// If "proof-of-work" feature is enabled
// assert!(package.check_proof_of_work(), "Invalid proof of work.");
println!("{:?}", package.public_data());

Receiving Package (with decrypting):

let mut package = encrypted_package::Package::
	<Vec<u8>, Vec<u8>, encrypted_package::DefaultConfig,
>::receive(&mut stream, None)?;
// If "proof-of-work" feature is enabled
// assert!(package.check_proof_of_work(), "Invalid proof of work.");
package.decrypt(&private_key)?;
assert!(package.check_signature()?, "Invalid signature.");
println!("{:?}\n{:?}", package.public_data(), package.data());

Re-exports

pub use config::Config;

pub use config::DefaultConfig;

Modules

config

All you need to configure encrypted packages.

consts

All consts for encrypted package.

Structs

Package

Encrypted package struct. PD is public data type, which doesn’t encrypt. D is private data type for decryption. C is config.

Enums

AesDecryptError

Error for AES decryption function.

AesEncryptError

Error for AES encryption function.

BuildDefaultConfigError

Error for build.

CheckSignatureError

Error for signature checking function.

CheckedSerializeError

Error for serialization with size check function.

DecryptError

Error for decrypt.

DefaultDefaultConfigError

Error for try_default.

GenerateRandomBytesError

Error for random bytes generation function.

NewError

Error for new.

ReceiveError

Error for receive and areceive if “async” feature is enabled.

SendError

Error for send and asend if “async” feature is enabled.

SignError

Error for sign.

ValidateSizeError

Error for size validation function.