Struct encrypted_package::Package

pub struct Package<PD, D, C> { /* private fields */ }

Encrypted package struct. PD is public data type, which doesn’t encrypt. D is private data type for decryption. C is config.

See the module level documentation and DefaultConfig for more.

Implementations

impl<PD, D, C> Package<PD, D, C>where
PD: DeserializeOwned + Serialize,
D: DeserializeOwned + Serialize,
C: Config + DeserializeOwned + Serialize,

pub fn new(
recipient_public_key: &Rsa<Public>,
public_data: Option<PD>,
data: D,
config: C
) -> Result<Self, NewError>

Creates a new encrypted package with config.

See also: DefaultConfig.

pub async fn areceive<R: AsyncReadExt + Unpin>(
r: &mut R,
size_limit: Option<usize>
) -> Result<Self, ReceiveError>

Receives bytes from a r with size_limit limit and deserializes into the structure.

See module level documentation for more algorithm details.

pub fn receive<R: Read>(
r: &mut R,
size_limit: Option<usize>
) -> Result<Self, ReceiveError>

Receives bytes from a r with size_limit limit and deserializes into the structure.

See module level documentation for more algorithm details.

pub fn config(&self) -> &C

Accessor for config.

pub fn data(&self) -> Option<&D>

Accessor for data.

pub fn public_data(&self) -> Option<&PD>

Accessor for public data.

pub async fn asend<W: AsyncWriteExt + Unpin>(
&self,
w: &mut W,
size_limit: Option<usize>
) -> Result<(), SendError>

Sends self to w.

See module level documentation for more algorithm details.

pub fn send<W: Write>(
&self,
w: &mut W,
size_limit: Option<usize>
) -> Result<(), SendError>

Sends self to w.

See module level documentation for more algorithm details.

pub fn decrypt(&mut self, private_key: &Rsa<Private>) -> Result<(), DecryptError>

With the private_key it decrypts data.

pub fn generate_proof_of_work(&mut self)

Generates a proof-of-work with difficulty self.config().proof_of_work_difficulty().

May take a long time. It is better to use this in conjunction with async blocking code spawner.

pub fn sign(
&mut self,
sender_private_key: &Rsa<Private>
) -> Result<(), SignError>

Signs the current hash with the private_key.

pub fn check_proof_of_work(&self) -> bool

Checks what current hash’s hex starts with self.config().proof_of_work_difficulty() zeros.

pub fn check_signature(&self) -> Result<bool, CheckSignatureError>

Checks that the sender has signed current hash with his private key.